third_party/WebKit/LayoutTests/http/tests/devtools/elements/edit/edit-dom-actions-shadow-1-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that user can mutate author shadow DOM by means of elements panel.


Running: testSetUp

Running: testSetAuthorShadowDOMElementAttribute
==== before ====
- <div id="testSetAuthorShadowDOMElementAttribute">
    - #shadow-root (open)
          <div foo="attribute value" id="shadow-node-to-set-attribute"></div>
  </div>
==== after ====
- <div id="testSetAuthorShadowDOMElementAttribute">
    - #shadow-root (open)
          <div id="shadow-node-to-set-attribute" bar="edited attribute"></div>
  </div>
